cleanup: remove base bitmap reader/writer from parquet module, use arrow bitutil ones
